Clicking tag calls foldByTags()
